reference analyzer
This is a vulnerable symbol reference analyzer
that is compatible with x/tools/analysis framework.

NOTE: This approach has limitations in handling
complex call graphs involving interfaces, types
with many methods, and can lead false positives.
For more precise analysis, use govulncheck.

This analyzer requires vulnerability information
loaded from Go vulndb (https://vuln.go.dev) as
mapping from package to a list of osv.Entry.
A vulnerability for a package can affect only
certain versions of the package. x/tools/analysis
does not carry the version information, thus,
this map should be populated carefully to include
only vulnerabilities that may affect the version of
each package to be analyzed.

A symbol appeared in the map (Catalog), it is vulnerable.

If a function references a vulnerable symbol, the function
symbol is vulnerable.

If a method references a vulnerable symbol, its receiver
type symbol is vulnerable. This heuristic can result in
false positives, but we want to be conservative.

If init function references a vulnerable symbol,
the package is vulnerable.

If a package imports a vulnerable package, the package
is vulnerable.

The analyzer exports info about exported vulnerable
symbols as object facts, and info about vulnerable
init functions as package facts.

TODO: include global variables; tainted global variables
make the package vulnerable like init functions.