vulns command that uses analysis.Analyzer
Unlike other traditional analyzers,
vuln/analysis.Analyzer requires Catalog
that carries vulnerable symbol information.
This command populates it, writes it in a temporary
file, and sets the analyzer's flag to point to
this temporary file that carries the Catalog
info.

Unlike other simple singlechecker, multichecker
based tools, this command uses a modified
checker API
(https://go-review.googlesource.com/c/tools/+/411907)
which allows to reuse the go/packages.Package
loaded for Catalog creation before, and to post-process
diagnostics results from analyzer.

When a project is affected a vulnerability, it
is possible that hundreds or more places end up
getting tainted. We are interested in false
positives/negatives of this approach. Hundreds
of diagnostic messages can be overwhelming.
We aggregate the diagnostic messages by
vulnerability ids. We preserve a single reference
path for each vulnerability so we can investigate
false positives/negatives.